requirements.txt
Django==1.6.5
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.9.6
argparse==1.2.1
dj-static==0.0.6
django-filter==0.7
djangorestframework==2.3.14
gunicorn==19.1.0
itsdangerous==0.24
markup==0.2
static3==0.5.1
